requirements.txt
pandas
sqlalchemy
python-dotenv
requests
psycopg2-binary
xlrd